clipp: clang (win) build ok, all tests passed
