+++ mk-msvs-solution-clang.cmd
@@ -6,11 +6,14 @@ rem set BUILD_CONF=Release
 set BUILD_CONF=Debug
 
 call %CD%\mk-vars.cmd
+call %CD%\mk-trace.cmd
 
 rem Compiler section --------------------------------------------------------------------
 set CLIPP_CMAKE_GENERATOR=%MSVC_CMAKE_GENERATOR%
-set CLIPP_C_COMPILER=%MSVC_C_COMPILER%
-set CLIPP_CXX_COMPILER=%MSVC_CXX_COMPILER%
+set CLIPP_C_COMPILER=%CLANG_C_COMPILER%
+set CLIPP_CXX_COMPILER=%CLANG_CXX_COMPILER%
+
+"%CLIPP_CXX_COMPILER%" --version
 
 rem -------------------------------------------------------------------------------------
 pushd %SOURCE_DIR%
@@ -22,7 +25,7 @@ rem ----------------------------------------------------------------------------
 rem -DCMAKE_TOOLCHAIN_FILE="%VCPKG_ROOT%\scripts\buildsystems\vcpkg.cmake" ^
 rem -DBLD_COMPUTER_NAME="%COMPUTER_NAME%" ^
 chcp.com 65001 >NUL
-cmake -G %CLIPP_CMAKE_GENERATOR% ^
+cmake -G %CLIPP_CMAKE_GENERATOR% -T ClangCL,host=x64 -A x64 ^
 -DCMAKE_BUILD_TYPE:STRING="%BUILD_CONF%" ^
 -DCMAKE_CONFIGURATION_TYPES:STRING="%BUILD_CONF%" ^
 -DCMAKE_C_COMPILER:FILEPATH="%CLIPP_C_COMPILER%" ^
