+++ mk-trace.cmd
@@ -1,0 +1,16 @@
+@echo off
+
+@echo            VCPKG_ROOT : %VCPKG_ROOT%
+@echo            BUILD_CONF : %BUILD_CONF%
+@echo            SOURCE_DIR : %SOURCE_DIR%
+@echo             BUILD_DIR : %BUILD_DIR%
+@echo             MSVC_ROOT : %MSVC_ROOT%
+@echo       MSVC_TOOLS_ROOT : %MSVC_TOOLS_ROOT%
+@echo         MSVC_KIT_ROOT : %MSVC_KIT_ROOT%
+@echo  MSVC_CMAKE_GENERATOR : %MSVC_CMAKE_GENERATOR%
+@echo       MSVC_C_COMPILER : %MSVC_C_COMPILER%
+@echo     MSVC_CXX_COMPILER : %MSVC_CXX_COMPILER%
+@echo CLANG_CMAKE_GENERATOR : %CLANG_CMAKE_GENERATOR%
+@echo      CLANG_C_COMPILER : %CLANG_C_COMPILER%
+@echo    CLANG_CXX_COMPILER : %CLANG_CXX_COMPILER%
+@echo           RELEASE_ENV : %RELEASE_ENV%
